requirements.txt
Django
Pillow
django-image-cropping
easy-thumbnails


